Relay Host-05's ping request to Host-06
It doesn't yet work. The ICMP echo request gets to Host-06 okay, but
Host-06 doesn't always send the reply back. I think it's because Host-06
doesn't get the ARP reply when it sends ARP request for the address of
Host-05. That's because Host-105 is spoofing Host-05's cache. So, even
if Host-05 sends the ARP reply to Host-06, it gets stuck at Host-105.
Because Host-105 doesn't relay ARP replies; as of yet it only relays
ICMP echo requests.