Use unsigned short & long for uint16_t & uint32_t
Apparently they are not the same. The uint*_t family are of type int.
And for some reason the TCP checksum function doesn't work with uint*_t
family. So when I tried to incorporate the general compute_checksum
function into tcp_checksum function, it did not work properly without
changing the variables used in compute_checksum to of unsigned short and
unsigned long type.

Also, the variable tcplen in tcp_checksum function needs to be of
unsigned short type, not int. There is something going on there, too,
that I don't know about. What a shitshow!